convert to typescript
Convert test and implementation to Typescript.
- add & configure ts-jest to run tests
- add a build script
- add & configure typescript & dependencies
- rename & enrich source files
