+++ package.json
@@ -2,8 +2,10 @@
   "name": "dotenv-defaults",
   "version": "0.0.0-development",
   "description": "dotenv... but with defaults!",
-  "main": "src/index.js",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
+    "build": "tsc -b",
     "test": "jest",
     "lint": "standard",
     "travis": "npm run lint && npm t"
@@ -33,11 +35,17 @@
     "dotenv": "^8.0.0"
   },
   "devDependencies": {
+    "@types/dotenv": "^6.1.1",
+    "@types/jest": "^24.0.15",
     "jest": "^24.8.0",
-    "standard": "^12.0.1"
+    "standard": "^12.0.1",
+    "ts-jest": "^24.0.2",
+    "typescript": "^3.5.2"
   },
   "jest": {
     "coverageDirectory": "./coverage/",
-    "collectCoverage": true
+    "collectCoverage": true,
+    "preset": "ts-jest",
+    "testEnvironment": "node"
   }
 }
